requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
groq 
python-dotenv
langchain-tavily
typing
typing_extensions
langgraph
langchain_community
groq